% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kreutz2023.clint.L3}
\alias{kreutz2023.clint.L3}
\alias{Kreutz2023.clint.L3}
\title{\insertCite{kreutz2023category;textual}{invitroTKdata} Intrinsic Hepatic Clearance Level-3 Data Set}
\format{
A level-3 data.frame with 25 rows and 13 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Calibration}}{Identifier for mass spectrometry calibration -- usually the date}
\item{\code{Clint}}{Frequentist point estimate for intrinsic hepatic clearance (Clint)}
\item{\code{Clint.pValue}}{p-value of the estimated \code{Clint} value}
\item{\code{Fit}}{Test nominal concentrations in the linear regression fit}
\item{\code{AIC}}{Akaike Information Criterion (AIC) for the linear regression fit}
\item{\code{AIC.Null}}{Akaike Information Criterion of the exponential decay assuming a constant rate of decay}
\item{\code{Clint.1}}{Intrinsic hepatic clearance at 1 uM (frequentist point estimate)}
\item{\code{Clint.10}}{Intrinsic hepatic clearance at 10 uM (frequentist point estimate)}
\item{\code{AIC.Sat}}{Akaike Information Criterion of the exponential decay with a saturation probability}
\item{\code{Sat.pValue}}{p-value of exponential decay with a saturation probability}
}
}
\usage{
kreutz2023.clint.L3
}
\description{
Frequentist estimates of intrinsic hepatic clearance (Clint) for
cryopreserved pooled human hepatocyte suspensions. Chemicals were per- and
polyfluoroalkyl substance (PFAS) samples. The experiments were
led by Dr.s Anna Kreutz and Barbara Wetmore.
}
\references{
\insertRef{shibata2002prediction}{invitroTKdata}

\insertRef{kreutz2023category}{invitroTKdata}
}
\keyword{datasets}
