% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kreutz2023.uc.L4}
\alias{kreutz2023.uc.L4}
\alias{Kreutz2023.uc.L4}
\title{\insertCite{kreutz2023category;textual}{invitroTKdata} Ultracentrifugation Level-4 Data Set}
\format{
A level-4 data.frame with 52 rows and 13 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Fstable.Med}}{Posterior median chemical stability fraction}
\item{\code{Fstable.Low}}{Posterior 2.5th quantile chemical stability fraction (lower credible interval bound)}
\item{\code{Fstable.High}}{Posterior 97.5th quantile chemical stability fraction (upper credible interval bound)}
\item{\code{Fup.Med}}{Posterior median fraction unbound in plasma}
\item{\code{Fup.Low}}{Posterior 2.5th quantile of fraction unbound in plasma (lower credible interval bound)}
\item{\code{Fup.High}}{Posterior 97.5th quantile of fraction unbound in plasma (upper credible interval bound)}
\item{\code{Fup.point}}{Point estimate of fraction unbound in plasma}
\item{\code{Unstable}}{Qualitative determination of chemical stability. "Y" indicates observed chemical stability.}
\item{\code{Uncertain}}{Qualitative determination of uncertainty about chemical stability. "Y" indicates uncertainty in observed chemical stability.}
\item{\code{CV}}{Coefficient of variance}
}
}
\usage{
kreutz2023.uc.L4
}
\description{
Bayesian estimates of plasma protein binding measured by
ultracentrifugation (UC) for per- and poly-fluorinated alkyl substance
(PFAS) samples from experiments led by Dr.s Anna Kreutz and Barbara Wetmore.
}
\references{
\insertRef{howard2010plasma}{invitroTKdata}

\insertRef{kreutz2023category}{invitroTKdata}
}
\keyword{data}
