% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smeltz2023.clint.L4}
\alias{smeltz2023.clint.L4}
\alias{Smeltz2023.clint.L4}
\title{\insertCite{smeltz2023plasma;textual}{invitroTKdata} Intrinsic Hepatic Clearance Level-4 Data Set}
\format{
A level-4 data.frame with 7 rows and 12 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Clint.1.Med}}{Posterior median intrinsic hepatic clearance at 1 uM}
\item{\code{Clint.1.Low}}{Posterior 2.5th quantile of intrinsic hepatic clearance at 1 uM (lower credible interval bound)}
\item{\code{Clint.1.High}}{Posterior 97.5th quantile of intrinsic hepatic clearance at 1 uM (upper credible interval bound)}
\item{\code{Clint.10.Med}}{Posterior median intrinsic hepatic clearance at 10 uM}
\item{\code{Clint.10.Low}}{Posterior 2.5th quantile of intrinsic hepatic clearance at 10 uM (lower credible interval bound)}
\item{\code{Clint.10.High}}{Posterior 97.5th quantile of intrinsic hepatic clearance at 10 uM (upper credible interval bound)}
\item{\code{Clint.pValue}}{Probability that a chemical concentration decrease is observed}
\item{\code{Sat.pValue}}{Probability that a lower \code{Clint} is observed at a higher concentration, i.e. saturation probability}
\item{\code{degrades.pValue}}{Probability of abiotic degradation}
}
}
\usage{
smeltz2023.clint.L4
}
\description{
Bayesian estimate of intrinsic hepatic clearance (Clint) for
cryopreserved pooled human hepatocytes. Chemicals were per- and
polyfluoroalkyl substance (PFAS) samples. The experiments were
led by Dr.s Marci Smeltz and Barbara Wetmore.
}
\references{
\insertRef{shibata2002prediction}{invitroTKdata}

\insertRef{smeltz2023plasma}{invitroTKdata}
}
\keyword{data}
