% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smeltz2023.uc.L3}
\alias{smeltz2023.uc.L3}
\alias{Smeltz2023.uc.L3}
\title{\insertCite{smeltz2023plasma;textual}{invitroTKdata} Ultracentrifugation Level-3 Data Set}
\format{
A level-3 data.frame with 107 rows and 5 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Calibration}}{Identifier for mass spectrometry calibration -- usually the date}
\item{\code{Fup}}{Frequentist point estimate for fraction unbound in plasma (fup)}
}
}
\usage{
smeltz2023.uc.L3
}
\description{
Frequentist estimate of plasma protein binding measured by
ultracentrifugation (UC) for per- and poly-fluorinated alkyl substance
(PFAS) samples from experiments led by Dr.s Marci Smeltz and Barbara Wetmore.
}
\references{
\insertRef{howard2010plasma}{invitroTKdata}

\insertRef{smeltz2023plasma}{invitroTKdata}
}
\keyword{data}
