% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smeltz2023.uc.L4}
\alias{smeltz2023.uc.L4}
\alias{Smeltz2023.uc.L4}
\title{\insertCite{smeltz2023plasma;textual}{invitroTKdata} Ultracentrifugation Level-4 Data Set}
\format{
A level-4 data.frame with 69 rows and 7 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Fup.Med}}{Posterior median fraction unbound in plasma}
\item{\code{Fup.Low}}{Posterior 2.5th quantile of fraction unbound in plasma (lower credible interval bound)}
\item{\code{Fup.High}}{Posterior 97.5th quantile of fraction unbound in plasma (upper credible interval bound)}
\item{\code{Fup.point}}{Point estimate of fraction unbound in plasma}
}
}
\usage{
smeltz2023.uc.L4
}
\description{
Bayesian estimate of plasma protein binding measured by
ultracentrifugation (UC) for per- and poly-fluorinated alkyl substance
(PFAS) samples from experiments led by Dr.s Marci Smeltz and Barbara Wetmore.
}
\references{
\insertRef{howard2010plasma}{invitroTKdata}

\insertRef{smeltz2023plasma}{invitroTKdata}
}
\keyword{data}
