% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.clint.L3}
\alias{wambaugh2019.clint.L3}
\title{\insertCite{wambaugh2019assessing;textual}{invitroTKdata} Intrinsic Hepatic Clearance Level-3 Data Set}
\format{
A data.frame with 473 rows and 13 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemical Dashboard)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Calibration}}{Identifier for mass spectrometry calibration - usually the date}
\item{\code{Clint}}{Frequentist point estimate for intrinsic hepatic clearance (Clint)}
\item{\code{Clint.pValue}}{p-value of the estimated \code{Clint} value}
\item{\code{Fit}}{Test nominal concentrations in the linear regression fit}
\item{\code{AIC}}{Akaike Information Criterion (AIC) for the linear regression fit}
\item{\code{AIC.Null}}{Akaike Information Criterion of the exponential decay assuming a constant rate of decay}
\item{\code{Clint.1}}{Intrinsic hepatic clearance at 1 uM (frequentist point estimate)}
\item{\code{Clint.10}}{Intrinsic hepatic clearance at 10 uM (frequentist point estimate)}
\item{\code{AIC.Sat}}{Akaike Information Criterion of the exponential decay with a saturation probability}
\item{\code{Sat.pValue}}{p-value of exponential decay with a saturation probability}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.clint.L3
}
\description{
Frequentist estimate of intrinsic hepatic clearance
(Clint) measured using \emph{in vitro} suspensions of pooled primary human
hepatocytes \insertCite{shibata2002prediction}{invitroTKdata}.
}
\references{
\insertRef{shibata2002prediction}{invitroTKdata}

\insertRef{wambaugh2019assessing}{invitroTKdata}
}
\keyword{data}
