% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.clint.L4}
\alias{wambaugh2019.clint.L4}
\title{\insertCite{wambaugh2019assessing;textual}{invitroTKdata} Intrinsic Hepatic Clearance Level-4 Data Set}
\format{
A level-4 data.frame with 473 rows and 12 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{Clint.1.Med}}{Posterior median intrinsic hepatic clearance at 1 uM}
\item{\code{Clint.1.Low}}{Posterior 2.5th quantile of intrinsic hepatic clearance at 1 uM (lower credible interval bound)}
\item{\code{Clint.1.High}}{Posterior 97.5th quantile of intrinsic hepatic clearance at 1 uM (upper credible interval bound)}
\item{\code{Clint.10.Med}}{Posterior median intrinsic hepatic clearance at 10 uM}
\item{\code{Clint.10.Low}}{Posterior 2.5th quantile of intrinsic hepatic clearance at 10 uM (lower credible interval bound)}
\item{\code{Clint.10.High}}{Posterior 97.5th quantile of intrinsic hepatic clearance at 10 uM (upper credible interval bound)}
\item{\code{Clint.pValue}}{Probability that a chemical concentration decrease is observed}
\item{\code{Sat.pValue}}{Probability that a lower \code{Clint} is observed at a higher concentration, i.e. saturation probability}
\item{\code{degrades.pValue}}{Probability of abiotic degradation}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.clint.L4
}
\description{
Bayesian estimate of intrinsic hepatic clearance
(Clint) measured using \emph{in vitro} suspensions of pooled primary human
hepatocytes \insertCite{shibata2002prediction}{invitroTKdata}.
}
\references{
\insertRef{shibata2002prediction}{invitroTKdata}

\insertRef{wambaugh2019assessing}{invitroTKdata}
}
\keyword{data}
