% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.red.L3}
\alias{wambaugh2019.red.L3}
\title{\insertCite{wambaugh2019assessing;textual}{invitroTKdata} Rapid Equilibrium Dialysis Level-3 Data Set}
\format{
A data.frame 368 rows and 4 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemicals Dashboard)}
\item{\code{Calibration}}{Identifier for mass spectrometry calibration - usually the date}
\item{\code{Fup}}{Frequentist point estimate for fraction unbound in plasma (fup)}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.red.L3
}
\description{
Frequentist estimate of plasma protein binding using the rapid equilibrium
dialysis (RED) assay method \insertCite{waters2008validation}{invitroTKdata}.
}
\references{
\insertRef{waters2008validation}{invitroTKdata}

\insertRef{wambaugh2019assessing}{invitroTKdata}
}
\keyword{data}
