% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.red.L4}
\alias{wambaugh2019.red.L4}
\title{\insertCite{wambaugh2019assessing;textual}{invitroTKdata} Rapid Equilibrium Dialysis Level-4 Data Set}
\format{
A data.frame 301 rows and 7 variables: \describe{
\item{\code{Compound.Name}}{Compound name}
\item{\code{Lab.Compound.Name}}{Compound as described in the laboratory}
\item{\code{DTXSID}}{DSSTox Substance Identifier (CompTox Chemical Dashboard)}
\item{\code{Fup.point}}{Point estimate of fraction unbound in plasma}
\item{\code{Fup.Med}}{Posterior median fraction unbound in plasma}
\item{\code{Fup.Low}}{Posterior 2.5th quantile of fraction unbound in plasma (lower credible interval bound)}
\item{\code{Fup.High}}{Posterior 97.5th quantile of fraction unbound in plasma (upper credible interval bound)}
}
}
\source{
Wambaugh et al. (2019)
}
\usage{
wambaugh2019.red.L4
}
\description{
Bayesian estimate of plasma protein binding using the rapid equilibrium
dialysis (RED) assay method \insertCite{waters2008validation}{invitroTKdata}.
}
\references{
\insertRef{waters2008validation}{invitroTKdata}

\insertRef{wambaugh2019assessing}{invitroTKdata}
}
\keyword{data}
